Remove getCredentialsFilename and getConfigFilename from UserCredentialsUtils
