+++ src/shared/credentials/userCredentialsUtils.ts
@@ -9,8 +9,8 @@ import * as path from 'path'
 import { writeFile } from 'fs-extra'
 import * as vscode from 'vscode'
 import * as nls from 'vscode-nls'
+import { getConfigFilename, getCredentialsFilename } from '../../credentials/sharedCredentials'
 import { credentialHelpUrl } from '../constants'
-import { EnvironmentVariables } from '../environmentVariables'
 import { mkdir } from '../filesystem'
 import { fileExists, readFileAsString } from '../filesystemUtilities'
 import { SystemUtilities } from '../systemUtilities'
@@ -41,7 +41,7 @@ export class UserCredentialsUtils {
      * @returns array of filenames for files found.
      */
     public static async findExistingCredentialsFilenames(): Promise<string[]> {
-        const candidateFiles: string[] = [this.getCredentialsFilename(), this.getConfigFilename()]
+        const candidateFiles: string[] = [getCredentialsFilename(), getConfigFilename()]
 
         const existsResults: boolean[] = await Promise.all(
             candidateFiles.map(async filename => await SystemUtilities.fileExists(filename))
@@ -50,33 +50,13 @@ export class UserCredentialsUtils {
         return candidateFiles.filter((filename, index) => existsResults[index])
     }
 
-    // TODO : CC : move to sharedCredentials.ts
-    /**
-     * @returns Filename for the credentials file
-     */
-    public static getCredentialsFilename(): string {
-        const env = process.env as EnvironmentVariables
-
-        return env.AWS_SHARED_CREDENTIALS_FILE || path.join(SystemUtilities.getHomeDirectory(), '.aws', 'credentials')
-    }
-
-    // TODO : CC : move to sharedCredentials.ts
-    /**
-     * @returns Filename for the config file
-     */
-    public static getConfigFilename(): string {
-        const env = process.env as EnvironmentVariables
-
-        return env.AWS_CONFIG_FILE || path.join(SystemUtilities.getHomeDirectory(), '.aws', 'config')
-    }
-
     /**
      * @description Determines if credentials directory exists
      * If it doesn't, creates credentials directory
-     * at directory from this.getCredentialsFilename()
+     * at directory from getCredentialsFilename()
      */
     public static async generateCredentialDirectoryIfNonexistent(): Promise<void> {
-        const filepath = path.dirname(this.getCredentialsFilename())
+        const filepath = path.dirname(getCredentialsFilename())
         if (!(await fileExists(filepath))) {
             await mkdir(filepath, { recursive: true })
         }
@@ -100,11 +80,11 @@ export class UserCredentialsUtils {
         const credentialsFileContents = handlebarTemplate(credentialsContext)
 
         // Make a final check
-        if (await SystemUtilities.fileExists(this.getCredentialsFilename())) {
+        if (await SystemUtilities.fileExists(getCredentialsFilename())) {
             throw new Error('Credentials file exists. Not overwriting it.')
         }
 
-        await writeFile(this.getCredentialsFilename(), credentialsFileContents, {
+        await writeFile(getCredentialsFilename(), credentialsFileContents, {
             encoding: 'utf8',
             mode: 0o100600 // basic file (type 100) with 600 permissions
         })
